Gpdviz as main object
