+++ .travis.yml
@@ -16,6 +16,6 @@ install:
 
 before_script:
 - psql -c 'create database gpdviz_test;' -U postgres
-- sbt clean "gpdvizJVM/runMain gpdviz.server.GpdvizServer generate-conf --overwrite"
+- sbt clean "gpdvizJVM/runMain gpdviz.Gpdviz generate-conf --overwrite"
 
 script: sbt test
